+++ src/SwiftCourier/Controllers/LocationsController.cs
@@ -5,20 +5,24 @@ using Microsoft.AspNet.Mvc.Rendering;
 using Microsoft.Data.Entity;
 using SwiftCourier.Models;
 using SwiftCourier.ViewModels;
+using Microsoft.AspNet.Identity;
 
 namespace SwiftCourier.Controllers
 {
     public class LocationsController : BaseController
     {
-        private ApplicationDbContext _context;
-
-        public LocationsController(ApplicationDbContext context)
+        public LocationsController(
+            UserManager<User> userManager, ApplicationDbContext context) : base(userManager, context)
         {
-            _context = context; 
         }
         
         public async Task<IActionResult> Index()
         {
+            if (!HasPermission("VIEW_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             var locations = await _context.Locations.ToListAsync();
 
             return View(locations.ToListViewModel());
@@ -26,6 +30,11 @@ namespace SwiftCourier.Controllers
         
         public async Task<IActionResult> Details(int? id)
         {
+            if (!HasPermission("VIEW_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             if (id == null)
             {
                 return HttpNotFound();
@@ -42,6 +51,11 @@ namespace SwiftCourier.Controllers
         
         public IActionResult Create()
         {
+            if (!HasPermission("CREATE_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             return View();
         }
         
@@ -49,6 +63,11 @@ namespace SwiftCourier.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create(LocationViewModel model)
         {
+            if (!HasPermission("CREATE_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             if (ModelState.IsValid)
             {
                 var location = model.ToEntity();
@@ -62,6 +81,11 @@ namespace SwiftCourier.Controllers
         
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!HasPermission("EDIT_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             if (id == null)
             {
                 return HttpNotFound();
@@ -79,6 +103,11 @@ namespace SwiftCourier.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(LocationViewModel model)
         {
+            if (!HasPermission("EDIT_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             if (ModelState.IsValid)
             {
                 var location = await _context.Locations.SingleAsync(m => m.Id == model.Id);
@@ -99,6 +128,11 @@ namespace SwiftCourier.Controllers
         [ActionName("Delete")]
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!HasPermission("DELETE_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             if (id == null)
             {
                 return HttpNotFound();
@@ -117,6 +151,11 @@ namespace SwiftCourier.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!HasPermission("DELETE_LOCATIONS"))
+            {
+                return Unauthorized();
+            }
+
             Location location = await _context.Locations.SingleAsync(m => m.Id == id);
             _context.Locations.Remove(location);
             await _context.SaveChangesAsync();
